Ensure wavelength indices are integer, not float

+++ ullyses/coadd.py
@@ -118,7 +118,7 @@ class SegmentList:
 
     def wavelength_to_index(self, wavelength):
         index = (wavelength - self.min_wavelength) / self.delta_wavelength
-        indices = [round(x) for x in index]
+        indices = [int(round(x)) for x in index]
         return indices
 
     def index_to_wavelength(self, index):
